Made the first step towards f64 support
The end goal is to allow users to switch between f64 and i64 registers. I now implemented a working solution. The code is not pretty, but it works, and that's the constrain right now.
What's left to do is to make the code more readable, mainly by hiding all the trait bounds away.